Add helper script and split Dockerfiles for faster dev
In this setup, the Docker image installs the Gem dependencies of the
theme, which takes some minutes. This is quite long time if you're
testing changes in the _config.yml and is not necessary. Making two
images, the second using the first, enables a rebuild that takes 10 secs
to test changes in _config.yml.