docs/develop: Replace g_dbus_proxy_call with g_dbus_connection_call
The example in "VAPIX access for ACAP applications" creates a
connection to the bus, creates a proxy object and then calls the method.
Creating a proxy for this use-case is unnecessary and does one or more
D-Bus calls under the hood as you do not pass
G_DBUS_PROXY_FLAGS_DO_NOT_LOAD_PROPERTIES which is necessary to make
this request "lightweight". The better solution is to skip the proxy
completely and just use g_dbus_connection_call_sync, as we do not use
any of the features of the proxy.

Change-Id: I1ba8781ab68492a426e150691b0015fc55be53d0